+++ zed_components/src/zed_camera/include/zed_camera_component.hpp
@@ -294,6 +294,7 @@ private:
   bool mPoseSmoothing = false;
   bool mAreaMemory = true;
   std::string mAreaMemoryDbPath = "";
+  sl::POSITIONAL_TRACKING_MODE mPosTrkMode = sl::POSITIONAL_TRACKING_MODE::QUALITY;
   bool mImuFusion = true;
   bool mFloorAlignment = false;
   bool mTwoDMode = false;