+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 LATEST CHANGES
 ==============
 
+2023-09-04
+----------
+- Add new parameter `pos_tracking.pos_tracking_mode` to exploit the new ZED SDK `QUALITY` mode for improved odometry and localization
+
 2023-09-03
 ----------
 - New Video/Depth processing throttling method by using the `grab_compute_capping_fps` ZED SDK parameter instead of a dedicated thread